contributors/gadams/incoming/cssom/cssstyledeclaration-inline-color.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD//XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>CSSOM - Inline Style - Serialization - Color</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="http://www.w3.org/TR/cssom/#the-cssstyledeclaration-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script><![CDATA[
      var style = document.getElementById('box').style;

      function trimFinalDelimiter(s) {
        s = s.trim();
        if ( ( s.length > 0 ) && ( s.lastIndexOf(";") == ( s.length - 1 ) ) ) {
          s = s.substring(0,s.length - 1);
        }
        return s;
      }

      test(function(){

        style.color = "red";
        assert_equals(trimFinalDelimiter(style.cssText), "color: red");

      }, 'set color \'red\', get inline text \'red\'');


      test(function(){

        style.color = "rgb(255,0,0)";
        assert_equals(trimFinalDelimiter(style.cssText), "color: rgb(255, 0, 0)");

      }, 'set color \'rgb(255,0,0)\', get inline text \'rgb(255, 0, 0)\'');
    ]]>
    </script>
  </body>
</html>
